refactor(client): consolidate duplicate types into game/types/
- Delete types/player.ts — its types were duplicated in game/types/
- Add quantity field to RawEquipmentSlot in game/types/equipment.ts
- Create game/types/events.ts as canonical NetworkEvents definition
- Update useModalPanels to import from game/types instead of local copy
- Re-route types/index.ts barrel to re-export from game/types/

Eliminates type duplication between types/player.ts and game/types/,
establishing game/types/ as the single source of truth for client
UI type definitions.
