+++ packages/client/src/game/types/equipment.ts
@@ -10,7 +10,11 @@
 import type { Item } from "@hyperscape/shared";
 
 /** Raw equipment slot format from server network cache */
-export type RawEquipmentSlot = { item: Item | null; itemId?: string } | null;
+export type RawEquipmentSlot = {
+  item: Item | null;
+  itemId?: string;
+  quantity?: number;
+} | null;
 
 /** Raw equipment data structure from server network cache */
 export type RawEquipmentData = {
