+++ packages/client/src/types/index.ts
@@ -51,14 +51,14 @@ export type {
   SelectOption,
 } from "./ui";
 
-// Player state types
+// Player state types (canonical definitions in game/types/)
 export type {
   RawEquipmentSlot,
   RawEquipmentData,
   InventorySlotViewItem,
   NetworkEventName,
-} from "./player";
-export { NetworkEvents } from "./player";
+} from "../game/types";
+export { NetworkEvents } from "../game/types";
 
 // Type guards for runtime validation
 export {
